// slider/noUiSlider/nouislider.d.ts
declare type Pips = PositionsPips | ValuesPips | CountPips | StepsPips | RangePips;
declare type StartValues = string | number | (string | number)[];
declare type HandleAttributes = {
  [key: string]: string;
};
interface UpdatableOptions {
  range?: Range;
  start?: StartValues;
  margin?: number;
  limit?: number;
  padding?: number | number[];
  snap?: boolean;
  step?: number;
  pips?: Pips;
  format?: Formatter;
  tooltips?: boolean | PartialFormatter | (boolean | PartialFormatter)[];
  animate?: boolean;
}
export interface Options extends UpdatableOptions {
  range: Range;
  connect?: 'lower' | 'upper' | boolean | boolean[];
  orientation?: 'vertical' | 'horizontal';
  direction?: 'ltr' | 'rtl';
  behaviour?: string;
  keyboardSupport?: boolean;
  keyboardPageMultiplier?: number;
  keyboardMultiplier?: number;
  keyboardDefaultStep?: number;
  documentElement?: HTMLElement;
  cssPrefix?: string;
  cssClasses?: CssClasses;
  ariaFormat?: PartialFormatter;
  animationDuration?: number;
  handleAttributes?: HandleAttributes[];
}
export interface API {
  destroy: () => void;
  steps: () => NextStepsForHandle[];
  on: (eventName: string, callback: EventCallback) => void;
  off: (eventName: string) => void;
  get: (unencoded?: boolean) => GetResult;
  set: (input: number | string | (number | string)[], fireSetEvent?: boolean, exactInput?: boolean) => void;
  setHandle: (handleNumber: number, value: number | string, fireSetEvent?: boolean, exactInput?: boolean) => void;
  reset: (fireSetEvent?: boolean) => void;
  options: Options;
  updateOptions: (optionsToUpdate: UpdatableOptions, fireSetEvent: boolean) => void;
  target: HTMLElement;
  removePips: () => void;
  removeTooltips: () => void;
  getTooltips: () => {
    [handleNumber: number]: HTMLElement | false;
  };
  getOrigins: () => {
    [handleNumber: number]: HTMLElement;
  };
  pips: (grid: Pips) => HTMLElement;
}
export interface TargetElement extends HTMLElement {
  noUiSlider?: API;
}
interface NearByStep {
  startValue: number;
  step: number | false;
  highestStep: number;
}
interface NearBySteps {
  stepBefore: NearByStep;
  thisStep: NearByStep;
  stepAfter: NearByStep;
}
declare type GetResult = number | string | (string | number)[];
declare type NextStepsForHandle = [number | false | null, number | false | null];
declare type PipsFilter = (value: number, type: PipsType) => PipsType;
declare type EventCallback = (this: API, values: (number | string)[], handleNumber: number, unencoded: number[], tap: boolean, locations: number[], slider: API) => void;